HTML page
This is an HTML page that has a working grid that changes colors. Nothing special. I wasn't able to link off the Javascript to an external page so if one of you want to do that go ahead. I will continue to try and work on the DNS and webserver so that it is working. This also has the new index.js page that I am using on the site. That is what I will launch when we make the page offically live so any mulitplayer functionality should be ran through that probs.